+++ source/js/search-parameters/observation-search-parameters.js
@@ -94,7 +94,10 @@ export const ObservationSearchParameters = () => ({
       if (selectedCodes.length === 1 && testAC.url.indexOf('&q=') === -1) {
         initTestAC(searchItemId, eventData.item_code, () => {
           testAC.domCache.set('elemVal', eventData.val_typed_in);
-          testAC.urlSearch(eventData.val_typed_in, Def.Autocompleter.Base.MAX_ITEMS_BELOW_FIELD);
+          testAC.urlSearch(
+            eventData.val_typed_in,
+            Def.Autocompleter.Base.MAX_ITEMS_BELOW_FIELD
+          );
         });
       } else if (
         selectedCodes.length === 0 &&